+++ apps/mis-server/src/utils/orm.ts
@@ -13,7 +13,9 @@
 import { Ref, Reference } from "@mikro-orm/core";
 import { EntityManager } from "@mikro-orm/mysql";
 
-export const paginationProps = (page?: number, pageSize: number = 10) => ({
+export const DEFAULT_PAGE_SIZE = 50;
+
+export const paginationProps = (page?: number, pageSize: number = DEFAULT_PAGE_SIZE) => ({
   offset: ((page ?? 1) - 1) * pageSize,
   limit: pageSize,
 });