+++ apps/mis-web/src/pageComponents/job/RunningJobTable.tsx
@@ -11,8 +11,9 @@
  */
 
 import { useDidUpdateEffect } from "@scow/lib-web/build/utils/hooks";
-import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/i18n";
-import { Button, Form, Input, InputNumber, Select, Space, Table } from "antd";
+import { DEFAULT_PAGE_SIZE } from "@scow/lib-web/build/utils/pagination";
+import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/systemLanguage";
+import { Button, Form, Input, InputNumber, message, Popconfirm, Select, Space, Table } from "antd";
 import React, { useCallback, useMemo, useRef, useState } from "react";
 import { useAsync } from "react-async";
 import { useStore } from "simstate";
@@ -250,7 +251,10 @@ export const RunningJobInfoTable: React.FC<JobInfoTableProps> = ({
         } : {})}
         dataSource={data}
         loading={isLoading}
-        pagination={{ showSizeChanger: true }}
+        pagination={{
+          showSizeChanger: true,
+          defaultPageSize: DEFAULT_PAGE_SIZE,
+        }}
         rowKey={runningJobId}
         scroll={{ x: data?.length ? 1800 : true }}
         tableLayout="fixed"
@@ -298,11 +302,27 @@ export const RunningJobInfoTable: React.FC<JobInfoTableProps> = ({
 
         <Table.Column<RunningJobInfo>
           title={t(pCommon("more"))}
-          width="9%"
+          width="12%"
           fixed="right"
           render={(_, r) => (
             <Space>
               <a onClick={() => setPreviewItem(r)}>{t(pCommon("detail"))}</a>
+              <Popconfirm
+                title={t(p("finishJobConfirm"))}
+                onConfirm={async () =>
+                  api.cancelJob({
+                    query: {
+                      cluster: r.cluster.id,
+                      jobId: r.jobId,
+                    },
+                  }).then(() => {
+                    message.success(t(p("finishJobSuccess")));
+                    reload();
+                  })
+                }
+              >
+                <a>{t(p("finishJobButton"))}</a>
+              </Popconfirm>
               <ChangeJobTimeLimitModalLink
                 reload={reload}
                 data={[r]}