+++ apps/portal-web/src/layouts/BaseLayout.tsx
@@ -13,7 +13,7 @@
 import { DatabaseOutlined } from "@ant-design/icons";
 import { BaseLayout as LibBaseLayout } from "@scow/lib-web/build/layouts/base/BaseLayout";
 import { JumpToAnotherLink } from "@scow/lib-web/build/layouts/base/header/components";
-import { PropsWithChildren, useMemo } from "react";
+import { PropsWithChildren } from "react";
 import { useStore } from "simstate";
 import { LanguageSwitcher } from "src/components/LanguageSwitcher";
 import { useI18n, useI18nTranslateToString } from "src/i18n";
@@ -26,9 +26,10 @@ import { publicConfig } from "src/utils/config";
 interface Props {
   footerText: string;
   versionTag: string | undefined;
+  initialLanguage: string;
 }
 
-export const BaseLayout = ({ footerText, versionTag, children }: PropsWithChildren<Props>) => {
+export const BaseLayout = ({ footerText, versionTag, initialLanguage, children }: PropsWithChildren<Props>) => {
 
   const userStore = useStore(UserStore);
   const { loginNodes } = useStore(LoginNodeStore);
@@ -37,9 +38,11 @@ export const BaseLayout = ({ footerText, versionTag, children }: PropsWithChildr
   const t = useI18nTranslateToString();
   const languageId = useI18n().currentLanguage.id;
 
-  const routes = useMemo(() => userRoutes(
+  const systemLanguageConfig = publicConfig.SYSTEM_LANGUAGE_CONFIG;
+
+  const routes = userRoutes(
     userStore.user, defaultCluster, loginNodes, setDefaultCluster,
-  ), [userStore.user, defaultCluster, setDefaultCluster]);
+  );
 
   const logout = () => {
     removeDefaultCluster();
@@ -64,7 +67,11 @@ export const BaseLayout = ({ footerText, versionTag, children }: PropsWithChildr
             link={publicConfig.MIS_URL}
             linkText={t("baseLayout.linkText")}
           />
-          <LanguageSwitcher />
+          {
+            systemLanguageConfig.isUsingI18n ? (
+              <LanguageSwitcher initialLanguage={initialLanguage} />
+            ) : undefined
+          }
         </>
       )}
     >