+++ apps/portal-web/src/pageComponents/filemanager/SingleCrossClusterTransferSelector.tsx
@@ -10,12 +10,12 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/i18n";
+import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/systemLanguage";
 import { Select } from "antd";
 import React, { useCallback } from "react";
 import { useAsync } from "react-async";
 import { api } from "src/apis";
-import { useI18n } from "src/i18n";
+import { prefix, useI18n, useI18nTranslateToString } from "src/i18n";
 import { Cluster } from "src/utils/config";
 
 
@@ -26,11 +26,14 @@ interface SingleSelectionProps {
   exclude?: Cluster;
 }
 
+const p = prefix("pageComp.fileManagerComp.singleCrossClusterTransferSelector.");
 
 export const SingleCrossClusterTransferSelector: React.FC<SingleSelectionProps> = ({
   value, onChange, label, exclude,
 }) => {
   const languageId = useI18n().currentLanguage.id;
+
+  const t = useI18nTranslateToString();
 
   const { data: availableClusters = [], isLoading, reload } = useAsync({
     promiseFn: useCallback(async () => {
@@ -43,7 +46,7 @@ export const SingleCrossClusterTransferSelector: React.FC<SingleSelectionProps> 
   return (
     <Select
       labelInValue
-      placeholder="请选择集群"
+      placeholder={t(p("placeholder"))}
       value={value ? ({ value: value.id, label: value.name }) : undefined}
       onChange={({ value, label }) => {
         onChange?.({ id: value, name: label });