+++ apps/portal-web/src/pages/shell/[cluster]/[loginNode]/[[...path]].tsx
@@ -12,7 +12,7 @@
 
 import "xterm/css/xterm.css";
 
-import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/i18n";
+import { getI18nConfigCurrentText } from "@scow/lib-web/build/utils/systemLanguage";
 import { Button, Popover, Space, Typography } from "antd";
 import { NextPage } from "next";
 import dynamic from "next/dynamic";