+++ dev/test-adapter/src/app.ts
@@ -13,9 +13,11 @@
 import { Server } from "@ddadaal/tsgrpc-server";
 import { config } from "src/config/env";
 import { accountServiceServer } from "src/services/account";
+import { appServiceServer } from "src/services/app";
 import { configServiceServer } from "src/services/config";
 import { jobServiceServer } from "src/services/job";
 import { userServiceServer } from "src/services/user";
+import { versionServiceServer } from "src/services/version";
 
 export async function createServer() {
 
@@ -35,6 +37,8 @@ export async function createServer() {
   await server.register(userServiceServer);
   await server.register(jobServiceServer);
   await server.register(configServiceServer);
+  await server.register(appServiceServer);
+  await server.register(versionServiceServer);
 
   return server;