+++ libs/scheduler-adapter/src/client.ts
@@ -12,9 +12,11 @@
 
 import { ChannelCredentials } from "@grpc/grpc-js";
 import { AccountServiceClient } from "@scow/scheduler-adapter-protos/build/protos/account";
+import { AppServiceClient } from "@scow/scheduler-adapter-protos/build/protos/app";
 import { ConfigServiceClient } from "@scow/scheduler-adapter-protos/build/protos/config";
 import { JobServiceClient } from "@scow/scheduler-adapter-protos/build/protos/job";
 import { UserServiceClient } from "@scow/scheduler-adapter-protos/build/protos/user";
+import { VersionServiceClient } from "@scow/scheduler-adapter-protos/build/protos/version";
 
 type ClientConstructor<TClient> =
   new (address: string, credentials: ChannelCredentials) => TClient;
@@ -24,6 +26,8 @@ export interface SchedulerAdapterClient {
   user: UserServiceClient;
   job: JobServiceClient;
   config: ConfigServiceClient;
+  version: VersionServiceClient;
+  app: AppServiceClient;
 }
 
 export function getClient<TClient>(
@@ -41,5 +45,7 @@ export const getSchedulerAdapterClient = (address: string) => {
     user: getClient(address, UserServiceClient),
     job: getClient(address, JobServiceClient),
     config: getClient(address, ConfigServiceClient),
+    version: getClient(address, VersionServiceClient),
+    app: getClient(address, AppServiceClient),
   };
 };